requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
PyYAML==6.0.1
pandas==2.1.4
openpyxl==3.1.2
Werkzeug==3.0.1
Jinja2==3.1.2
python-dateutil==2.8.2
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
